refactor: Refactor pytest integration
